% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{openbeta_server.path}
\alias{openbeta_server.path}
\title{Path of OpenBeta Server}
\usage{
openbeta_server.path()
}
\value{
Existing directory path
}
\description{
Fails if the OpenBeta Server product has not been installed.
}
\examples{
library(dcsr)

# List all files with extension `.miz` found recursively under the OpenBeta
# Server folder. Escape the full stop and match the end of the file name. Name
# the vector of paths by the base name.
file.miz <- dcsr::openbeta_server.path() \%>\%
  list.files(pattern = "\\\\.miz$", full.names = TRUE, recursive = TRUE)
names(file.miz) <- basename(file.miz)

# Are any duplicated by base name? Not at the time of writing.
basename(file.miz) \%>\%
  duplicated() \%>\%
  any()
#> [1] FALSE

# Extract one of the missions. Unzipping as raw produces a list of raw
# vectors without encoding. The mission payload contains Russian text in the
# following example.
fw190.miz <- file.miz["fw190.miz"] \%>\%
  dcsr::unzip_bin()
fw190.miz \%>\% str()
#> List of 3
#>  $ mission   : raw [1:26985] 6d 69 73 73 ...
#>  $ warehouses: raw [1:16514] 77 61 72 65 ...
#>  $ options   : raw [1:3448] 6f 70 74 69 ...
mission <- rawToChar(fw190.miz$mission)
Encoding(mission) <- "UTF-8"
}
