% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unz.R
\name{unzip_bin}
\alias{unzip_bin}
\title{Unzip File to Raw Vectors}
\usage{
unzip_bin(zipfile)
}
\arguments{
\item{zipfile}{Path of zip file}
}
\value{
List of raw vectors by zip member name
}
\description{
Automatically filters out members with file names ending with a forward
slash. Trailing slash denotes a sub-directory.
}
